Break markdown rendering into module sections
Move markdown object rendering from fulldoc setup into a dedicated module template with explicit sections, so users can override parts like YARD's default templates.\n\nFixes #5.
